Agregue la logica del TileSet.cs y refactorize el TileRenderer.cs, tambien Arregle un bug visual en Camera2D haciendo cast a numeros enteros

+++ src/Aurora.Engine/Camera/Camera2D.cs
@@ -8,8 +8,8 @@ public class Camera2D
 
     public Matrix Transform => 
         Matrix.CreateTranslation(
-            -Position.X,
-            -Position.Y,
+            -(int)Position.X,
+            -(int)Position.Y,
             0f
         );
 }
